Reduces the granularity of transactions within Notify Handler
- refactors models.Database to provide a Connection
- models.Connection provides a Transaction that can be created wherever
- transaction was moved to wrap only the postal.Mailer Deliver function

[#78075920]

+++ postal/user_loader_test.go
@@ -26,7 +26,7 @@ var _ = Describe("UserLoader", func() {
 
             tokenClaims := map[string]interface{}{
                 "client_id": "mister-client",
-                "exp":       3404281214,
+                "exp":       int64(3404281214),
                 "scope":     []string{"notifications.write"},
             }
